--- src/For_training/train1.java
@@ -1,12 +1,0 @@
-package For_training;
-
-public class train1 {
-
-    String name = "OzCompany";
-
-
-
-
-
-
-}
