+++ src/forTraining/Cat.java
@@ -1,4 +1,4 @@
-package For_training;
+package forTraining;
 
 public class Cat {
