+++ src/forTraining/Train1.java
@@ -1,0 +1,17 @@
+package forTraining;
+
+public class Train1 {
+
+    private String [] names;
+
+   public Train1(String [] names)  {
+     this.names = names;
+   }
+
+   public String [] getNames() {
+       return this.names;
+   }
+
+
+
+}
