+++ src/module2/FifthClass.java
@@ -1,8 +1,8 @@
-package Module2;
+package module2;
 
-public class SecondClass {
+public class FifthClass {
     public static void main(String[] args) {
         System.out.println("Мой первый класс на Java");
-        System.out.println("second");
+
     }
 }
