+++ src/module2/FirstClass.java
@@ -1,7 +1,8 @@
-package Module2;
+package module2;
 
-public class ThirdClass {
+public class FirstClass {
     public static void main(String[] args) {
         System.out.println("Мой первый класс на Java");
+        System.out.println("Hi");
     }
 }
