+++ src/module2/FourthClass.java
@@ -1,8 +1,7 @@
-package Module2;
+package module2;
 
-public class FifthClass {
+public class FourthClass {
     public static void main(String[] args) {
         System.out.println("Мой первый класс на Java");
-
     }
 }
