+++ src/module2/ThirdClass.java
@@ -1,8 +1,7 @@
-package Module2;
+package module2;
 
-public class FirstClass {
+public class ThirdClass {
     public static void main(String[] args) {
         System.out.println("Мой первый класс на Java");
-        System.out.println("Hi");
     }
 }
