+++ src/module3_HW/SpecialStudent.java
@@ -1,4 +1,4 @@
-package Module3_HW;
+package module3_HW;
 
 public class SpecialStudent extends CollegeStuden {
     private long secretKey;
